requirements.txt
openai==0.27.4
pyyaml==6.0
google-api-python-client
atlassian-python-api==3.36.0
flask
numpy
orjson